fix: rename captureHandler to captureRouter
