+++ server/app.js
@@ -8,7 +8,7 @@ const helper = require("./routes/utils");
 
 const app = express();
 const config = require('../config/config.js');
-const captureRouter = require('./routes/captureHandler');
+const captureRouter = require('./routes/captureRouter');
 const registerEventHandlers = require('./services/EventHandlers');
 
 Sentry.init({ dsn: config.sentry_dsn });
